Move FlightLog.lua to moduels/FlightLog/FlightLog
Do move first then we'll start splitting it up to keep git history nice.
