+++ data/pigui/modules/info-view/06-flightlog.lua
@@ -4,7 +4,7 @@
 local ui = require 'pigui'
 local InfoView = require 'pigui.views.info-view'
 local Lang = require 'Lang'
-local FlightLog = require 'FlightLog'
+local FlightLog = require 'modules.FlightLog.FlightLog'
 local Format = require 'Format'
 local Color = _G.Color
 local Vector2 = _G.Vector2
