+++ data/pigui/modules/new-game-window/class.lua
@@ -8,7 +8,7 @@ local msgbox = require 'pigui.libs.message-box'
 local Character = require 'Character'
 local Commodities = require 'Commodities'
 local Equipment = require 'Equipment'
-local FlightLog = require 'FlightLog'
+local FlightLog = require 'modules.FlightLog.FlightLog'
 local ModalWindow = require 'pigui.libs.modal-win'
 local ModelSkin = require 'SceneGraph.ModelSkin'
 local ShipDef = require "ShipDef"
