+++ data/pigui/views/mainmenu.lua
@@ -8,7 +8,7 @@ local ShipDef = require 'ShipDef'
 local Equipment = require 'Equipment'
 local MusicPlayer = require 'modules.MusicPlayer'
 local Lang = require 'Lang'
-local FlightLog = require 'FlightLog'
+local FlightLog = require 'modules.FlightLog.FlightLog'
 local Character = require 'Character'
 local Vector2 = _G.Vector2
 local NewGameWindow = require("pigui.modules.new-game-window.class")
